requirements.txt
aiohttp~=3.10.5
asyncpraw==7.7.1
edge_tts==6.1.12
faster_whisper==1.0.3
google_api_python_client==2.142.0
google_auth_oauthlib==1.2.1
librosa~=0.10.0
moviepy==1.0.3
mysql_connector_repackaged==0.3.1
numpy~=1.22.0
opencv_python==4.9.0.80
protobuf==5.28.0
python-dotenv==1.0.1
mysql~=0.0.3
mysql-connector-python~=9.0.0
opencv-python~=4.10.0.84